RAG Chatbot with Langchain/requirements.txt
langchain
langchain_community
langchain_core
python-dotenv
streamlit
faiss-cpu
pypdf
langchain-ollama
